+++ jaco_arm/jaco_gazebo/src/JacoGazeboJointStatePublisher.cpp
@@ -80,8 +80,8 @@ void JacoGazeboJointStatePublisher::Load(physics::ModelPtr _parent, sdf::Element
     // get joint names from parameters
     std::vector<std::string> joint_names;
     joints.getJointNames(joint_names, true);
-    const std::vector<double>& arm_init = joints.getArmJointsInitPose();
-    const std::vector<double>& finger_init = joints.getFingerJointsInitPose();
+    const std::vector<float>& arm_init = joints.getArmJointsInitPose();
+    const std::vector<float>& finger_init = joints.getFingerJointsInitPose();
 
     if (joint_names.size() != 9)
     {
